+++ src/ru/GB/LVL1/lesson1/MainApp.java
@@ -1,0 +1,7 @@
+package ru.GB.LVL1.lesson1;
+
+public class MainApp {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
