+++ src/ru/GB/LVL1/lesson3/Lesson3.java
@@ -1,0 +1,7 @@
+package ru.GB.LVL1.lesson3;
+
+public class Lesson3 {
+    public static void main(String[] args) {
+
+    }
+}
