+++ src/ru/GB/LVL2/Lesson1/Lesson1.java
@@ -1,0 +1,14 @@
+package ru.GB.LVL2.Lesson1;
+
+public class Lesson1 {
+    private int i=0;
+    private static int a;
+
+    public static void main(String[] args) {
+        System.out.println("1");
+
+    }
+
+
+
+}
