+++ src/ru/GB/LVL2/Lesson1/homework/RunLine.java
@@ -1,0 +1,13 @@
+package ru.GB.LVL2.Lesson1.homework;
+
+public class RunLine {
+    int dist;
+
+    public int getDist() {
+        return dist;
+    }
+
+    public void setDist(int dist) {
+        this.dist = dist;
+    }
+}
