+++ src/ru/GB/LVL2/Lesson1/homework/Wall.java
@@ -1,0 +1,13 @@
+package ru.GB.LVL2.Lesson1.homework;
+
+public class Wall {
+    int height;
+
+    public int getHeight() {
+        return height;
+    }
+
+    public void setHeight(int height) {
+        this.height = height;
+    }
+}
