+++ src/ru/GB/LVL2/Lesson2/homework/MyArraySizeException.java
@@ -1,0 +1,5 @@
+package ru.GB.LVL2.Lesson2.homework;
+
+public class MyArraySizeException extends Exception {
+
+}
