+++ src/ru/GB/LVL2/Lesson2Gusev/Ap.java
@@ -1,0 +1,12 @@
+package ru.GB.LVL2.Lesson2Gusev;
+
+public class Ap {
+
+    public static void main(String[] args) {
+        User user = new User(01);
+        Guard guard = new Guard();
+        boolean canPass = guard.canPass(user);
+
+    }
+
+}
