+++ src/ru/GB/LVL2/Lesson2Gusev/ComputerUser.java
@@ -1,0 +1,9 @@
+package ru.GB.LVL2.Lesson2Gusev;
+
+public class ComputerUser extends User {
+    private int iq;
+
+    public ComputerUser(int id) {
+        super(id);
+    }
+}
