+++ src/ru/GB/LVL2/Lesson2Gusev/Guard.java
@@ -1,0 +1,11 @@
+package ru.GB.LVL2.Lesson2Gusev;
+
+public class Guard {
+
+    boolean canPass(User user){
+        return user.getAgeBefore18() == 0;
+    };
+
+
+
+}
