--- src/ru/geekbrains/lesson1/MainApp.java
@@ -1,7 +1,0 @@
-package ru.geekbrains.lesson1;
-
-public class MainApp {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");
-    }
-}
