--- src/ru/geekbrains/lesson3/Lesson3.java
@@ -1,7 +1,0 @@
-package ru.geekbrains.lesson3;
-
-public class Lesson3 {
-    public static void main(String[] args) {
-
-    }
-}
