--- src/ru/geekbrainsLev2/Lesson1/Lesson1.java
@@ -1,11 +1,0 @@
-package ru.geekbrainsLev2.Lesson1;
-
-public class Lesson1 {
-    private int i=0;
-    private static int a;
-
-    class Inner {};
-
-
-
-}
