--- src/ru/geekbrainsLev2/Lesson1/homework/ru/geekbrainsLev2/Lesson1/homework/RunLine.java
@@ -1,13 +1,0 @@
-package ru.geekbrainsLev2.Lesson1.homework;
-
-public class RunLine {
-    int dist;
-
-    public int getDist() {
-        return dist;
-    }
-
-    public void setDist(int dist) {
-        this.dist = dist;
-    }
-}
