--- src/ru/geekbrainsLev2/Lesson1/homework/ru/geekbrainsLev2/Lesson1/homework/Wall.java
@@ -1,13 +1,0 @@
-package ru.geekbrainsLev2.Lesson1.homework;
-
-public class Wall {
-    int height;
-
-    public int getHeight() {
-        return height;
-    }
-
-    public void setHeight(int height) {
-        this.height = height;
-    }
-}
